// src/app/models/ConceptRelReq.model.ts
export interface ConceptRelReq{
    idConceptUp: number;
    idApplication: number;
    idCompany: number;
    username: string;
    nomCatalog: string;
    idConceptDown: number;
    idRelation: number;
    indValue: number;
    indNivel: number;
    operationType: number;    
    attribute1: string;
}